# Feature/matrix flattening