~35% faster grid updating because we don't have to dereference twice (row, then column).

```
Before:

running 3 tests
test grid_1000x1000_advance_10_times ... bench:  46,540,732 ns/iter (+/- 9,468,205)
test grid_500x500_advance_10_times   ... bench:  13,165,339 ns/iter (+/- 1,746,717)
test grid_50x50_advance_50_times     ... bench:   2,309,584 ns/iter (+/- 521,540)

After:
running 3 tests
test grid_1000x1000_advance_10_times ... bench:  30,145,033 ns/iter (+/- 1,775,850)
test grid_500x500_advance_10_times   ... bench:   8,656,586 ns/iter (+/- 1,088,516)
test grid_50x50_advance_50_times     ... bench:   1,411,911 ns/iter (+/- 262,043)
```